Add post box type to post box preset
I've added a single combo-key for post_box:type with values pillar, lamp, wall & meter.

I've excluded "post_box:type=community" which appears to be a tagging confusion with boxes for receiving post particularly in North America. As these are common in CH I'd expect there are more appropriate tags for these. Various versions of indoor don't seem to have consensus (and in the UK these are merely a specific pillar box design), so I've excluded those, and all tag values in CJK languages.

Meter as a separate value for a pillar box for pre-paid post has always seemed a bit odd, and it may make sense to exclude it if it is highly UK specific. The main three values certainly apply in a wide range of countries (IR,GB,FR,ES,DE,CH,PL,IT,AR to my certain knowledge).

First attempt to modify anything here, so hope this works :-).